refactor moved entity products from package controller to package entities
