+++ src/main/java/com/br/dorotech/app/repositories/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.br.dorotech.app.repositories;
 
-import com.br.dorotech.app.controllers.Products;
+import com.br.dorotech.app.models.entities.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Long, Products> {
